customisable colorscheme via -background and -foreground
This commit also changes the default colorscheme from black
text on a white background to white text on a black background,
which is more canonical for a screensaver and potentially better
at avoiding screen burn-in.

The old behaviour can be gotten back by changing the screen hack
command to: unicode -root -background white -foreground black
in the xscreensaver or gnome-screensaver configuration.